Move cart related components to new library structure
