+++ projects/storefrontlib/src/cms-components/checkout/cart/add-to-cart/add-to-cart.module.ts
@@ -1,15 +1,14 @@
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
 import {
   CmsConfig,
   ConfigModule,
-  UrlTranslationModule,
   I18nModule,
+  UrlTranslationModule,
 } from '@spartacus/core';
-
+import { SpinnerModule } from '../../../../lib/ui/components/spinner/spinner.module';
 import { CartSharedModule } from './../cart-shared/cart-shared.module';
-import { SpinnerModule } from './../../ui/components/spinner/spinner.module';
 import { AddToCartComponent } from './add-to-cart.component';
 import { AddedToCartDialogComponent } from './added-to-cart-dialog/added-to-cart-dialog.component';
