+++ projects/storefrontlib/src/cms-components/checkout/cart/add-to-cart/added-to-cart-dialog/added-to-cart-dialog.component.scss
