+++ projects/storefrontlib/src/cms-components/checkout/cart/cart-shared/cart-item-list/cart-item-list.component.scss
