+++ projects/storefrontlib/src/cms-components/checkout/cart/cart-shared/cart-item/cart-item.component.html
