+++ projects/storefrontlib/src/cms-components/checkout/index.ts
@@ -1,0 +1,1 @@
+export * from './cart/index';
