--- projects/storefrontlib/src/lib/cart/index.ts
@@ -1,8 +1,0 @@
-export * from './cart.module';
-export * from './cart-details/cart-details.module';
-export * from './add-to-cart/add-to-cart.module';
-export * from './cart-shared/cart-item/cart-item.component';
-export * from './cart-details/cart-details.component';
-export * from './cart-totals/cart-totals.component';
-export * from './cart-shared/order-summary/order-summary.component';
-export * from './add-to-cart/add-to-cart.component';
