+++ projects/storefrontlib/src/lib/checkout/checkout.module.ts
@@ -1,10 +1,10 @@
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { MultiStepCheckoutModule } from './components/multi-step-checkout/multi-step-checkout.module';
-import { CartComponentModule } from './../cart/cart.module';
-
-import { guards } from './guards/index';
+import { NgModule } from '@angular/core';
 import { CheckoutModule } from '@spartacus/core';
+import { CartComponentModule } from '../../cms-components/checkout/cart/cart.module';
+import { MultiStepCheckoutModule } from './components/multi-step-checkout/multi-step-checkout.module';
+import { guards } from './guards/index';
+
 @NgModule({
   imports: [
     CommonModule,
