+++ projects/storefrontlib/src/lib/checkout/components/multi-step-checkout/multi-step-checkout.module.ts
@@ -1,20 +1,19 @@
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
-
 import {
-  UrlTranslationModule,
-  ConfigModule,
-  CmsConfig,
   CheckoutModule,
+  CmsConfig,
+  ConfigModule,
   I18nModule,
+  UrlTranslationModule,
 } from '@spartacus/core';
-import { ShippingAddressModule } from './shipping-address/shipping-address.module';
+import { CartSharedModule } from '../../../../cms-components/checkout/cart/cart-shared/cart-shared.module';
+import { MultiStepCheckoutComponent } from './container/multi-step-checkout.component';
 import { DeliveryModeModule } from './delivery-mode/delivery-mode.module';
 import { PaymentMethodModule } from './payment-method/payment-method.module';
 import { ReviewSubmitModule } from './review-submit/review-submit.module';
-import { MultiStepCheckoutComponent } from './container/multi-step-checkout.component';
-import { CartSharedModule } from '../../../cart/cart-shared/cart-shared.module';
+import { ShippingAddressModule } from './shipping-address/shipping-address.module';
 
 @NgModule({
   imports: [
