+++ projects/storefrontlib/src/lib/checkout/components/multi-step-checkout/review-submit/review-submit.module.ts
@@ -1,10 +1,9 @@
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-
+import { NgModule } from '@angular/core';
 import { I18nModule } from '@spartacus/core';
+import { CartSharedModule } from '../../../../../cms-components/checkout/cart/cart-shared/cart-shared.module';
 import { CardModule } from '../../../../ui/components/card/card.module';
 import { ReviewSubmitComponent } from './review-submit.component';
-import { CartSharedModule } from '../../../../cart/cart-shared/cart-shared.module';
 
 @NgModule({
   imports: [CommonModule, CardModule, CartSharedModule, I18nModule],
