+++ projects/storefrontlib/src/lib/cms-lib/cms-lib.module.ts
@@ -1,9 +1,8 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
+import { CartComponentModule } from '../../cms-components/checkout';
 import { SiteContextSelectorModule } from '../../cms-components/misc/site-context-selector/site-context-selector.module';
 import { HamburgerMenuModule, SkipLinkModule } from '../../layout/index';
-import { CartDetailsModule } from '../cart/cart-details/cart-details.module';
-import { CartTotalsModule } from '../cart/cart-totals/cart-totals.module';
 import { OrderDetailsModule } from '../my-account/order/order-details/order-details.module';
 import { OrderHistoryModule } from '../my-account/order/order-history/order-history.module';
 import { PaymentMethodsModule } from '../my-account/payment-methods/payment-methods.module';
@@ -18,7 +17,6 @@ import { BreadcrumbModule } from './breadcrumb/breadcrumb.module';
 import { CategoryNavigationModule } from './category-navigation/category-navigation.module';
 import { FooterNavigationModule } from './footer-navigation/footer-navigation.module';
 import { LinkModule } from './link/link.module';
-import { MiniCartModule } from './mini-cart/mini-cart.module';
 import { NavigationModule } from './navigation/navigation.module';
 import { CmsParagraphModule } from './paragraph/paragraph.module';
 import { ProductCarouselModule } from './product-carousel/product-carousel.module';
@@ -38,19 +36,17 @@ import { SearchBoxModule } from './search-box/search-box.module';
     BreadcrumbModule,
     ProductCarouselModule,
     SearchBoxModule,
-    MiniCartModule,
     SiteContextSelectorModule,
     AddressBookModule,
     OrderHistoryModule,
     ProductListModule,
     ProductTabsModule,
-    CartDetailsModule,
-    CartTotalsModule,
     OrderDetailsModule,
     PaymentMethodsModule,
     UpdateEmailModule,
     UpdatePasswordModule,
     UpdateProfileModule,
+    CartComponentModule,
   ],
 })
 export class CmsLibModule {}
