+++ projects/storefrontlib/src/lib/product/components/product-details/product-details.module.ts
@@ -1,18 +1,17 @@
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { RouterModule } from '@angular/router';
-
 import { I18nModule } from '@spartacus/core';
-import { ProductDetailsComponent } from './container/product-details.component';
-import { ProductSummaryComponent } from './product-summary/product-summary.component';
-import { ProductImagesComponent } from './product-images/product-images.component';
-import { CmsModule } from '../../../cms/cms.module'; // some slots are loaded inside components (i.e. tabs)
+import { AddToCartModule } from '../../../../cms-components/checkout';
 // guards
-import { CartSharedModule } from './../../../cart/cart-shared/cart-shared.module';
-import { ComponentsModule } from './../../../ui/components/components.module';
-import { AddToCartModule } from '../../../cart/add-to-cart/add-to-cart.module';
+import { CartSharedModule } from '../../../../cms-components/checkout/cart/cart-shared/cart-shared.module';
+import { CmsModule } from '../../../cms/cms.module'; // some slots are loaded inside components (i.e. tabs)
 import { OutletModule } from '../../../outlet/index';
+import { ComponentsModule } from './../../../ui/components/components.module';
+import { ProductDetailsComponent } from './container/product-details.component';
+import { ProductImagesComponent } from './product-images/product-images.component';
+import { ProductSummaryComponent } from './product-summary/product-summary.component';
 
 @NgModule({
   imports: [
