+++ projects/storefrontlib/src/lib/product/components/product-tabs/product-tabs.module.ts
@@ -1,27 +1,25 @@
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { RouterModule } from '@angular/router';
-
 import {
   CmsConfig,
   ConfigModule,
+  I18nModule,
   ProductService,
   RoutingService,
   WindowRef,
-  I18nModule,
 } from '@spartacus/core';
-import { CmsModule } from '../../../cms/cms.module'; // some slots are loaded inside components (i.e. tabs)
 // guards
-import { CartSharedModule } from './../../../cart/cart-shared/cart-shared.module';
-import { ComponentsModule } from './../../../ui/components/components.module';
-import { AddToCartModule } from '../../../cart/add-to-cart/add-to-cart.module';
+import { CartSharedModule } from '../../../../cms-components/checkout/cart/cart-shared/cart-shared.module';
+import { PageComponentModule } from '../../../../cms-structure/page/component/page-component.module';
+import { CmsModule } from '../../../cms/cms.module'; // some slots are loaded inside components (i.e. tabs)
 import { OutletModule } from '../../../outlet/index';
+import { ComponentsModule } from './../../../ui/components/components.module';
 import { ProductTabsComponent } from './container/product-tabs.component';
 import { ProductAttributesComponent } from './product-attributes/product-attributes.component';
 import { ProductReviewsComponent } from './product-reviews/product-reviews.component';
 import { ProductReviewsModule } from './product-reviews/product-reviews.module';
-import { PageComponentModule } from '../../../../cms-structure/page/component/page-component.module';
 
 @NgModule({
   imports: [
@@ -32,7 +30,6 @@ import { PageComponentModule } from '../../../../cms-structure/page/component/pa
     ComponentsModule,
     CartSharedModule,
     CmsModule,
-    AddToCartModule,
     OutletModule,
     ProductReviewsModule,
     PageComponentModule,
