+++ projects/storefrontlib/src/public_api.ts
@@ -5,9 +5,9 @@
 export * from './cms-components/index';
 export * from './cms-structure/index';
 export * from './layout/index';
-export * from './lib/cart/index';
 export * from './lib/checkout/index';
 export * from './lib/cms-lib/index';
+export * from './lib/cms-route/index';
 export * from './lib/cms/index';
 export * from './lib/global-message/index';
 export * from './lib/my-account/index';
@@ -18,8 +18,7 @@ export * from './lib/pwa/pwa.module';
 export * from './lib/store-finder/index';
 export * from './lib/storefront-config';
 export * from './lib/storefront.module';
-export * from './lib/ui/index';
-export * from './lib/cms-route/index';
 export * from './lib/suffix-routes/index';
+export * from './lib/ui/index';
 export * from './lib/user/index';
 export * from './translations/index';
